Fix crash in 'bfv_statistics' test with enable_mergejoin=on.
The test forcibly modifies pg_statistics.stavalues, with a array that
doesn't match the column's datatype. In general, if you muck around the
catalogs like that, all bets are off, but since we have a test like that
in the test suite, let's at least make it work.

The problem only arises with enable_mergejoin=on, because apparently the
cost estimation of other join types follow a different codepath that
happens to work.