Fix assertion failure in cost estimation of merge joins.
When I run the regression tests with enable_mergejoin=on, I got:

+FATAL:  Unexpected internal error (costsize.c:2010)
+DETAIL:  FailedAssertion("!(outer_skip_rows <= outer_rows)", File: "costsize.c", Line: 2010)
+HINT:  Process 26232 will wait for gp_debug_linger=120 seconds before termination.
+Note that its locks and other resources will not be released until then.
+server closed the connection unexpectedly
+	This probably means the server terminated abnormally
+	before or while processing the request.
+connection to server was lost

That happened in the 'gp_recursive_cte' test, with this query:

select recursive_table_1.id from recursive_table_1, recursive_table_2 where recursive_table_1.id = recursive_table_2.id and EXISTS (select * from r where r.i = recursive_table_2.id);

I tried to reduce that to a simpler test case, but gave up. I'm sure it
could be done, but I wasn't able to, with quick testing. This bug seems
unlikely to reappear in the same form, and it's covered by the existing
test, even if it's a bit accidental, so that's good enough for me.