Fix missing_xlog with longer timeout
There is 5 sec delay between mirror retry to connect to primary. The original
timeout to detect for the streaming state change was 10 seconds, which caused
stability issue of the test if the probe happen to fall between two mirror
retries.

The fix increase the timeout to 20 seconds (200), which is also consistent with
rest of the timeout used in the test.

Author: Xin Zhang <xzhang@pivotal.io>