Fix a case with a Hash Agg and ORDER BY.
A Hash Agg doesn't preserve sort order. The grouping extensions planner
didn't get the memo. That caused an assertion failure, when a Gather Motion
with a Merge Key was put on top of a Hash Agg, without a Sort, and the rows
didn't come out in the expected order.

This fixes the assertion fault mentioned in github issue #4488, but not
the main issue of incorrect results.