Remove tests on default_storage_parameters GUC in different contexts.
It's a GUC, like any other. Testing that it's set correctly in every
possible context, when set to a user, when set on a database, etc., seems
excessive.

Also, if by some freak incident the GUC is set incorrectly, the
consequences are not very serious. The table just gets created with wrong
parameters, but it's still a valid table, and if you look at it with e.g.
\d, you will see the options actually used.