Remove workarounds for bug in ancient macOS version
The referenced bug does not affect modern versions of macOS (tested
back to El Capitan), and Snow Leopard is EOL since long.  Further,
this workaround is not present/needed in PostgreSQL, with the code
running on Leopard as well as Snow Leopard in the buildfarm without
issues, so it seems that this problem has been fixed even in Snow
Leopard in the later patchlevels.  Remove the workarounds and align
the code with upstream.