Fix handling of NULL datum in CDefaultComparator
Consider the query below

```
create table test_country (a char(40));
select * from test_country where a not in (NULL,'');
```

`CDefaultComparator` is not able to handle NULL datum in above case.

Comparison in cases of text/char dataypes are converted into comparator expression
with result being a result expression like below.

```
+--CScalarCmp (>)
   |--CScalarConst (161096236.000)
   +--CScalarConst (161104428.000)

Result Expression:

+--CScalarConst (0)
```

The result expression is either a 0 or 1 CScalarConstant. But in case of NULL datums,
the comparator always evaluates to NULL.

```
+--CScalarCmp (>)
        |--CScalarConst (161096236.000)
        +--CScalarConst (null)

Result Expression:

 +--CScalarConst (null)
```

This commits handles the case of NULL datum by assuming :
1. NULL is less than everything not NULL.
2. NULL is equal to NULL.