Bump statement_mem to 250MB for ORCA CI
commit a7cce539025a6105151bc2f3d4fe9929da941063 first bumped
statement_mem for ICW to 250MB, to address the issue of gpcheckcat
needing more memory in the later stages of installcheck-world.
installcheck-good, didn't really need the extra memory back in July
2017. However, memory consumption has slowly crept up in the past half a
year and now we have to bump this up or some or the tests will fail
under planner, duh.