Delete transaction_management/mpp16087_race_condition test.
The test is very random, as just runs a combination of same SQL
statements concurrently for 100 iterations and then checks if no
catalog issues have been reported. And funny catalog check if CI
actually happens before running those iterations.

Checked the history to see the bug for which the test was added and
its in very core piece of DTM, so ideally any flaw there would show-up
in ICW parallel test runs Or otherwise more targeted and specific test
must be witten for the same.