Delete transaction_management/high_transaction_id test.
Based on history the test was added for bug where a debug log would
cause buffer overflow and corrupt global memory in production when
transaction id is 9 characters or more. That code is long gone, plus
any such issues should be reported by coverity and fixed via same. The
current test is pretty heavy as performs pg_resetxlog to bump
transaction-id on each and every primary and mirror to have 9 digits.