Add error code to error, to avoid printing source file name and lineno.
Internal errors, i.e. errors without an error code, have the source file
name and line number appended to the error message. That's troublesome for
regression tests, as the line number needs to be memorized in the expected
output. Use a proper error code, to avoid that, and because it's nice to
have proper error codes anyway.