Remove residual FIXME from window merge work
During window merge, DQA was temporarily disabled with GPORCA.
After the following PR, this was resolved. In this PR, I am cleaning a
residual outdated fixme

https://github.com/greenplum-db/gpdb/pull/4086

The string_agg function was categorized as ordered window aggregate in
GPDB 5. So GPORCA did not support that and fallbacks to planner. In GPDB
6 string_agg is a straightforward window aggregate, so GPORCA supports
it.