Reorder gp_toolkit test
gp_toolkit tests runs a query which checks if there are any GUCs
which have different value on the cluster node. However, test
udf_exception_blocks_panic_scenarios is executed before it and it
injects PANIC which causes recovery and it takes some time for all
the GUCs to be in consistent state between the segments. But by that
time gp_toolkit identifies difference and fails, so moving gp_toolkit
test above it.