Prevent EXPLAIN DXL to return output in case of error
Commit ca4c29d496a9b3eefda6e6f287c3d31f3e4bb0f6 removed the error
deferring from EXPLAIN such that it attempted to return results
even in case of errors, but missed the DXL case. This removes the
TRY/CATCH wraps from DXL EXPLAIN to match the rest of the EXPLAIN
code.

Also remove the last (found) callsite that expects that a palloc
for a statsbuffer can have returned NULL into an Assertion.