Update CONTEXT: lines in expected output.
They are ignored by gpdiff, but it's nice to keep them up-to-date anyway,
because if a test fails, these differences alre also printed in
regression.diffs, even though they're not causing the failure.

There are some more CONTEXTs that in more complicated cases that are not
always the same from run to run, but this is a good start.