Reindent inline functions in tuplestorenew
The inline functions were single statement functions written on a
single line, which breaks the unittest mocker code when preprocessor
directives are added after these functions. The reason for this is
that the mocker code hunts for an ending brace on its own line. As a
later commit will introduce said preprocessor directive, it's time
to bite the bullet and fix this.