Fix SyncRepWaitForLSN() to ignore QueryCancelPending
Under MPP setting, we cannot allow primary go out of sync with mirror
because that will cause the segment inconsistent with rest of segments
in the cluster.

Hence instead of honoring the query cancel request and mirror not be in
sync with primary, the primary will actually ignore the query cancel
request after primary commits.

If such out of sync caused by mirror failure, regular FTS failure
handling will handle it.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>
Co-authored-by: Xin Zhang <xzhang@pivotal.io>