Make dispatching of pg_highest_oid() more robust.
Add more checks to make sure it returns a valid OID. Turn an assertion
into an elog(), so that you get a graceful error rather than crash if
assertions are not enabled. Schema-qualify the call, so that we don't
get fooled by a UDF called pg_highest_oid().

These errors are "can't happen" scenarios, but we've nevertheless seen the
assertion fail occasionally in the pipeline. I believe the failure has been
caused by bugs elsewhere, and these changes won't fix it, but it doesn't
hurt to check for errors more thoroughly here.