Error out when Metadata Names contain unicode characters incompatible w.r.t system's locale
All the metadata objects in ORCA are serialized and deserialized as wide
character strings (CWStringDynamic). The metadata name which is in char*
format; needs to be first converted into a wide character format before
it can be serialized. To perform this function, ORCA uses vswprintf()
library function. This function is sensitive to locale LC_CTYPE. On
Linux platforms, when the locale of the host system is set to "C" and
the metadata name contains unicode string, vswprintf() fails with return
value as -1. ORCA assumes that the failure happen due to insufficient
size of destination buffer and thus performs multiple retries with
increased buffer size. The function fails at each attempt due to
incorrect locale. Eventually ORCA crashes when it runs out of memory.

This commit checks the error code when vswprintf() fails and detects the
locale incompatibility. We fall back to planner in such cases.

This commit also adds a unit test that runs only on non Darwin
platforms.

Signed-off-by: Dhanashree Kashid <dkashid@pivotal.io>