Mirror should not get marked as DOWN after primary recovery.
Added additional global timestamp to track the end of recovery where
database server is ready to accept connections in variable
PMAcceptingConnectionsStartTime.

During checking mirror status, we compute the grace period not only
based on when last wal sender died, but also based on when recovery
process finished.

This will fix the issue where the mirror is marked down due to long
recovery time on the primary.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>
Co-authored-by: Xin Zhang <xzhang@pivotal.io>