Whitespace and formatting fixes.
The immediate reason to do this was the "this ‘else’ clause does not guard"
gcc warning from create_mergejoin_path(). But while we're at it, might as
well clean up the whole file.

I spotted one piece of code that looks broken, marked that with a FIXME to
make sure we revisit that.