Make test case less sensitive to number of databases and roles.
The number of databases and roles in the system affected the join order
that the planner chose. I sometimes saw failures because of that, depending
on what other tests I had run before. Add WHERE clauses on database and
role names, to fix the row-count estimate on the SeqScans to 1, regardless
of how many databases and roles there are.