Fix 'btree_gist' contrib module.
The btree_gist contrib module was originally removed from GPDB. It was not
present in the original GPDB code dump when GPDB was open sourced. However,
somewhere along the line, as part of the merges from upstream, the code was
re-imported from the upstream. Except for the Makefile and the regression
tests.

Having the .c files, but no Makefile and no tests, clearly doesn't make
sense. Bring back the Makefile and the tests from the upstream, so that the
whole module is in the same condition as in the upstream.

Fix the 'cash' regression test to not rely on OIDs, because OIDs on user
tables work differently in GPDB. With that change, and by adding an
'init_file' for pg_regress, all the regression tests pass.

I'm not sure if we want to start compiling and running the tests as part of
the usual build and "make installcheck-world", so I didn't do that yet. But
at least what we have in the repository now makes sense and works, if you
do compile it.