Remove SYNC_REP_DISABLED state
This state was meant to prevent auxiliary processes from waiting from syncrep.
As far as we can tell, auxiliary process should not be calling
SyncRepWaitForLSN() unless it is committing a transaction.  And if it is
committing a transaction, it better wait for mirror to flush XLOG upto the
commit LSN.  One example of an auxiliary process committing a transaction is
FTS, when it updates cluster configuration on master.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>