bfv_aggregate 8.4 MERGE fixme
Remove the 8.4 merge fixme for window function related tests. Before the merge,
ORCA and planner would both error out if the aggregate used in the window did
not have preliminary functions associated with it. After the merge, ORCA still
falls back to the planner, but the planner produces the correct plan.