Remove GPDB_90_MERGE_FIXME from hstore.sql
Prior to GPORCA version 2.55.10, ORCA produced a plan with Redistribute
Motion on the hstore attribute(which is not a "GPDB hashable" type).
This caused queries to fail with the error like:
ERROR:  Type 40963 is not hashable.
This FIXME was added in `contrib/hstore/sql/hstore.sql` to run such
failing queries with planner. Since the issue has been fixed in ORCA,
removing it.