Remove GPDB_84_MERGE_FIXME from outfuncs.c
The pseudocols field was added to RangeTblEntry to avoid duplicates in the
target list. When the field was added, a WRITE_NODE_FIELD was added in
_outRangeTblEntry(), but a corresponding READ_NODE_FIELD was not added in the
_readRangeTblEntry(). Additionally, this field was not added to the readfast
and outfast functions either since this field was not used after the planning
stage and did not have to be dispatched to the QE's.  Later, in
fd6741f97ac1f02143541314db4e0ebfe89f9815, a corresponding READ_NODE_FIELD was
added along with a FIXME to decide whether or not to serialize/deserialize.
After confirming that this field is not used after planning, we have decided to
remove any serialization/deserialization of the pseudocols member of
RangeTblEntry
The pseudocols member of RangeTblEntry is a list of CdbRelColumnInfo. This
structure was added in the same original commit. An _outCdbRelColumnInfo() was
added for this structure as well. Similarly, no corresponding
_readCdbRelColumnInfo() was added. This data structure is only used for making
pseudocols, so we have also removed the _outCdbRelColumnInfo() function.

Signed-off-by: Abhijit Subramanya <asubramanya@pivotal.io>