Remove GPDB_84_MERGE_FIXME from postmaster.c
The FIXME comment asked whether or not we wanted to remove the check
on CAC_WAITBACKUP.  It is true that we do not currently use WAITBACKUP
state, but we should continue to have the code.  It will prevent some
merge conflicts during postgres merge, and we may use it sometime in
the future if applicable to any future Greenplum features.

[ci skip]