Add unit test and re-organize code for serialization of COptimizerConfig
Previously, there was no unit test for the parsing or the serialization of a
COptimizerConfig object. This commit adds this test.

This commit also introduces some re-organization of production code. The
re-organization was motivated by the desire to clean up the call hierarchy for
the serialization of the COptimizerConfig class.  The entry point for
serializing a plan into DXL is COptimizer::PdxlnOptimize. If the minidump
traceflag was set, it will initialize a minidump and serialize the various
components needed. Serialization of the COptimizerConfig is done at the end as
part of CMinidumperUtils::Finalize, while adding the Error Context.

In the course of adding this test, it became clear that, though
CDXLUtils::SerializeOptimizerConfig was being called from
CSerializeOptimizerConfig::Serialize, our "production" code, in this case, it
was bending over backwards to accommodate the need to add traceflags to the
same XML serializer while still implementing the CSerializable interface for
the Serialize function.  As such, we have moved the logic of this function into
a Serialize function called directly on a COptimizerConfig object, which is
called in production from CSerializableOptimizerConfig::Serialize and in test
from CParseHandlerTestUtils::SerializeOptimizerConfig.  It should be noted that
this test does not activate the parse handler for a COptimizerConfig directly
and instead relies on the fact that a COptimizerConfig object is nested only
one level below the top-level DXLMessage tag. Thus, this function calls
CDXLUtils::PoptimizerConfigParseDXL which calls PphdxlParseDXL and uses the
poconf member.

Signed-off-by: Abhijit Subramanya <asubramanya@pivotal.io>