Bump number of retries in missing_xlog test.
Failures are seen in CI randomly where
checkpoint_and_wait_for_replication_replay return false, but when check on the
box later the primaries and mirrors are in sync. So, seems its timing related
failures maybe sometimes containers are running slow causing delays. This test
is for checking functionality and not performance, so avoiding flaky failures is
better with increased number of retries, which in general shouldn't affect
runtime.