Serialize nestloop_factor inside CostParams.
Previously we did not serialize the cost parameter nestloop factor and hence it
was missing from the minidump files. When tests are run through ctest, they are
unable to read the nestloop factor from the minidump files and pick up the
default value of nestloop factor which is 1. This causes a lot of the tests to
fail. Hence we would like to serialize the nestloop factor so that when a
minidump file is added for testing purposes, it contains the desired nestloop
factor. This commit serializes nestloop factor inside CostParams and also adds
a unit test to check the serialization and deserialization of the CostParams.

Signed-off-by: Sambitesh Dash <sdash@pivotal.io>