Don't pass superuser flag of SessionUserId/OuterUserId to segments
GUC "is_supersuer" only provide value for SHOW to display, it's
useless on segments. Those two flags are all designed to determine
the value of is_superuser, so it's unnecessary to pass them to the
segments.

With this commit, another problem is resolved, GPDB used to dispatch
a command within empty transaction and resource owner to define an
index concurrently. However, a syscache access in superuser_arg()
may report a SIGSEGV because CurrentResourceOwner is NULL, this commit use
SessionUserIsSuperuser instead of superuser_arg() to avoid such error.