Add GUC to enable / disable Join Associativity in ORCA
This commit introduces a GUC `optimizer_enable_associativity` to enable
or disable join associativity. Join Associativity increases the search
space as it increases the numbers of groups to represent a join and its
associative counterpart, i.e (A X B) X C ~ A X (B X C).

This patch, by default disables join associativity transform, if
required the users can enable the transform. There are few plan changes
which are observed due to this change. However, further evaluation of
the plan changes revealed that even though the cost of the the resulting
plan has increased, the execution time went down by 1-2 seconds.

For the queries with plan changes, there are 3 tables which are joined,
i.e A, B and C. If we increase the number of tuples returned by the
subquery which forms A', we see the old plan. But if the tuples in
relation B and C is significantly higher, the plan changes with the
patch yeild faster execution times. This suggests that we may need to
tune the cost model to adapt to such cases.

The plan cost increase is 1000x as compared to the old plans, this 1000x
factor is due to the value of `optimizer_nestloop_factor=1024`, if you
set the value of the GUC `optimizer_nestloop_factor=1`, the plan before
or after the patch remains same.