Fix gpdeletesystem tablespace query
In a recent commit, spclocation field from pg_tablespace catalog table
was removed. A query in gpdeletesystem uses this field and was missed
in that effort. We fix the issue by using the new
gp_tablespace_location() catalog function to find the user-created
tablespace locations on each host.

Reference to spclocation removal:
https://github.com/greenplum-db/gpdb/commit/4483b7d3bc16fd8b78076543f1ae696d267944b7