Metadata tracking for ALTER ROLE IN DATABASE
There was a GPDB_90_MERGE_FIXME regarding what we should report for
ALTER ROLE IN DATABASE in pg_stat_last_shoperation table for metadata
tracking. We should always report that the role object was altered if
coming from ALTER ROLE code path. To do this, we just check if the
role oid is valid because the call to AlterSetting from ALTER DATABASE
uses InvalidOid for role oid. While we're at it, we might as well add
some tests for pg_stat_last_shoperation.