Change SLES12 disk size in master pipeline
With the switch to GCP, the AWS variables no longer do anything. This
made the SLES12 VMs have default 64GB disk size and caused ICW
failures due to a recently introduced test that fills up the disk
pretty quick. This commit changes the disk size back to 128GB for GCP
as a quick fix while we investigate the disk consumption of the
culprit test.

[ci skip]