ci: Remove staggered starts
The staggered starting was necessary when the performance of concourse
was improved by limiting the spike in started jobs at any one time.
Performance is not as much a driving factor with concourse changes and
with reduction in size of the pipeline. Therefore the staggered starting
idea can be retired.

Authored-by: Kris Macoskey <kmacoskey@pivotal.io>