Pull start_prepare fault before HOLD_INTERRUPTS.
Sometimes test need to use this fault to suspend and then stop the database. If
this fault is inside the HOLD_INTERRUPTS (which is where its before this
change), CHECK_FOR_INTERRUPTS doesn't do anything. Hence stopping of segment
takes super long time as kill -9 needs to be enforced by gpstop after timeout.

Co-authored-by: David Kimura <dkimura@pivotal.io>