TINC: Delete pg_twophase test and CI job.
Decent coverage exist for 2PC in isolation2, plus coverage will be enhanced. The
current tests in TINC for pg_twophase are written specifically for filerep. That
much extensive tests to create all types of objects and all is not required for
walrep given the fact it relies on crash recovery logic for replication. Either
these tests need to be modified to continue to work for walrep, instead better
to delete them and write fresh ones as required for walrep and 2PC interaction
in isolation2.