Readers should check abort status of their subxids
QE readers incorrectly return true for TransactionIdIsCurrentTransactionId()
when passed with an xid that is an aborted subtransaction of current
transaction.  The end effect is wrong results because tuples inserted by the
aborted subtransaction are seen (treated as visible according to MVCC rules) by
a reader.  Current patch fixes the bug by looking up abort status of an XID
from pg_clog.  In a QE writer, just like in upstream PostgreSQL, subtransaction
information is available in CurrentTransactionState (even when subxip cache has
overflown).  This information is not maintained in shared memory, making it
unavailable to a reader.  Readers must resort to a longer route to get the same
information - pg_subtrans and pg_clog.

The patch does not use TransactionIdDidAbort() to check abort status.  The
interface is designed to work with all transaction IDs.  It walks up the
transaction hierarchy to look for an aborted parent if status of the given
transaction is found to be SUB_COMMITTED.  This is a wasted effort when a QE
reader wants to test if its own subtransaction has aborted or not.  A new
interface is introduced to avoid this wasted effort for QE readers.  We choose
to rely on AbortSubTransaction()'s behavior to mark entire subtree under the
aborted subtransaction to be aborted in pg_clog.  A SUB_COMMITTED status in
pg_clog, therefore, allows us to conclude that the subtransaction is not
aborted without having to walk up the hierarchy, provided, the subtransaction
is child of our own transaction.

The test case also needed a fix because the SQL query (insert into select *)
didn't result in a reader gang being created.  The SQL is changed to a join on
a non-distribution column so as to achieve reader gang creation.