Add missing Xforms that generate IndexedNLJ to PbsIndexJoinXforms()
PbsIndexJoinXforms() returns a set of xforms that generate indexed
nested loop join. The xforms added where not added to this function when
they where created, this causes xforms to be fired when the user wants
to disable all xforms that generate indexed nested loop join

Also bump ORCA version to 2.56.2.