Eliminate distributed transaction log creation and maintenance on QD.
Commit b3f300b9457 eliminated consulting distributed log in
XidInMVCCSnapshot(). This was based on in the QD, the distributed transactions
become visible at the same time as the corresponding local ones, so we can rely
on the local XIDs only. So, given QD never consults distributed transaction log,
lets completely eliminate its creation and maintenance on QD.

Note, during initdb identity (QD or QE) is currently not known and hence still
32k initial zero-filled distributed log file gets created on QD.