Refactor I - Remove fDisjunctionOrNegation
For a subquery that appears inside an OR or NOT expression in the WHERE
clause (EsqctxtFilter context), SubqueryHandler uses flag,
fDisjunctionOrNegation such that the subsequent subquery removal
routines treat the subquery as if it appeared in value context.

The current commit simplifies this mechanism by changing the context
from `EsqctxtFilter` to `EsqctxtValue` early on and thus eliminating the
need of passing fDisjunctionOrNegation around.

Signed-off-by: Dhanashree Kashid <dkashid@pivotal.io>