Treat non-scalar subq inside a scalar-expr to be in value context
A non-scalar subquery nested in a scalar expression that appears in a
predicate, is incorrectly treated to be always in a filtering context by
FRecursiveHandler. For the query mentioned below, the ANY subquery
nested inside a ScalarCmp was treated in a filtering context as the
expression appeared in WHERE clause, resulting into an APPLY as new
logical child and a constant true as residual scalar.  Hence after
subquery-apply transformation the original scalar comparison got
rewritten as “b = 1” resulting in a wrong plan.

Select2Apply xform Input:

+--CLogicalSelect
   |--CLogicalGet "t1" ("t1"), Columns: ["a" (0), "b" (1),..]
   +--CScalarCmp (=)
      |--CScalarIdent "b" (1)
      +--CScalarSubqueryAny(=)["a" (9)]
         |--CLogicalGet "t2" ("t2"), Columns: ["a" (9), ..]
         +--CScalarIdent "a" (0)
Output:
+--CLogicalLeftSemiApplyIn (Reqd Inner Cols: "a" (9))
   |--CLogicalSelect
   |  |--CLogicalGet "t1" ("t1"), Columns: ["a" (0), "b" (1), ..]
   |  +--CScalarCmp (=)
   |     |--CScalarIdent "b" (1)
   |     +--CScalarConst (1)
   |--CLogicalSelect
   |  |--CLogicalGet "t2" ("t2"), Columns: ["a" (9),..]
   |  +--CScalarCmp (=)
   |     |--CScalarIdent "a" (0)
   |     +--CScalarIdent "a" (9)
   +--CScalarConst (1)

However in such cases, the non-scalar subquery should be treated in
value context such that we generate an outer-apply expression with a
subquery column returned as residual scalar. This column identifier is
used to replace the subquery node in the parent tree.

Signed-off-by: Dhanashree Kashid <dkashid@pivotal.io>