gpaddmirrors: changes for WAL replication
Prior to this commit, gpaddmirrors was missing two bits of work which
were previously done by gpinitsystem/gpcreateseg. When adding mirrors to
a cluster, the pg_hba.conf on primary segments has to be modified to
allow replication connections, e.g. for pgbasebackup. And after the
mirrors are built, the catalog has to be updated to reflect the new
cluster config.

Co-authored-by: Nadeem Ghani <nghani@pivotal.io>
Co-authored-by: Jim Doty <jdoty@pivotal.io>