Correctly set pg_exttable.logerrors (#4985)
Consider the following SQL, we expect logging to be turned off for table
`ext_error_logging_off`

```sql
create external table ext_error_logging_off (a int, b int)
    location ('file:///tmp/test.txt') format 'text'
    segment reject limit 100;
\d+ ext_error_logging_off
```
And then in this next case we expect error logging to be turned on for
table `ext_t2`:

```sql
create external table ext_t2 (a int, b int)
    location ('file:///tmp/test.txt') format 'text'
    log errors segment reject limit 100;
\d+ ext_t2
```

Before this patch, we are making two mistakes in handling these external
table DDL:

1. We intend to enable error logging *whenever* the user specifies
`SEGMENT REJECT` clause, completely ignoring whether he or she specifies
`LOG ERRORS` 1. Even then, we make the mistake of implicitly coercing
the OID (an unsigned 32-bit integer) to a bool (which is really just a C
`char`): that means, 255/256 of the time (99.6%) the result is `true`,
and 0.4% of the time we get a `false` instead.

The `OID` to `bool` implicit conversion could have been caught by a
`-Wconversion` GCC/Clang flag. It's most likely a leftover from commit
8f6fe2d60b6d7e93d35cca2d80cc360ad8732599.

This bug manifests itself in the `dsp` regression test mysteriously
failing about once every 200 runs -- with the only diff on a `\d+` of an
external table that should have error logging turned on, but the
returned definition has it turned off.

While working on this we discovered that all of our existing external
tables have both `LOG ERRORS` and `SEGMENT REJECT`, which is why this
bug wasn't caught in the first place.

This patch fixes the issue by properly setting the catalog column
`pg_exttable.logerrors` according to the user input.

While we were at it, we also cleaned up a few dead pieces of code and
made the `dsp` test a bit friendlier to debug.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>
Co-authored-by: David Kimura <dkimura@pivotal.io>