Fix how we print xform results
When we set the `gucs optimizer_print_xform` and
`optimizer_print_xform_result`, the intent is to see what a specific
transform has an input and see what output it produces if there is one.

The way the transforms currently work is that we apply a specific
transform to all possible trees rooted from a specific group expression,
and we store all the results in one object called `CXformResult`.  When
we print the result of an xfrom for a specifc tree rooted from the same
group expression, we print all previous results as well. This is very
misleading as if all the alternatives are created applying the transform
on that specific tree.

This commit fixes this issue and prints an input and than the actual
produced output rather than accumulated result where a specific
transform is applied to the trees rooted by a specific group expression.

Let's say that group expression `GROUPEXPRE1` can produce two possible
expression trees. `EXPR1` and `EXPR2`. And we apply `XFORM1` on these
expression trees. And the output expressions are `OUTEXPR1` and
`OUTEXPR2`.

Previosuly the output of the tranform would look like below:

```
--XFORM1--
Input:
	EXPR1
Output:
Alternatives:
0:
	OUTEXPR1

--XFORM1--
Input:
	EXPR2
Output:
Alternatives:
0:
	OUTEXPR1
1:
	OUTEXPR2
```

One would easily think that the second application of the transform
`EXPR2` produced 2 alternatives which is misleading and wrong.

After the fix:

```
--XFORM1--
Input:
	EXPR1
Output:
Alternatives:
0:
	OUTEXPR1

--XFORM1--
Input:
	EXPR2
Output:
Alternatives:
0:
	OUTEXPR2
```