Always generate relfilenode for about to be created relations
We have found the culprit causing relfilenode collisions to be VACUUM
FULL on a mapped relation.  The code was reusing OID as relfilenode for
the temporary table created by vacuum full.  This happened without
bumping the relfilenode counter.  The patch fixes this such that
relfilenode is always generated, even in case of mapped relations.

With this, we believe that the possibility of collision still exists in the way
sequence OIDs are generated.  That needs to be fixed in a separate patch.  The
fixme in GetNewRelFileNode() should be sufficient to note this.

Co-authored-by: David Kimura <dkimura@pivotal.io>