Add ServerAliveInterval=60 option to ssh.
For long running commands such as gpinitstandby with a large master data
directory, the server takes a long time. Therefore, there is no acitivity from
the client to the server. If the ClientAliveInterval is set, then the server
reports a timeout after ClientAliveInterval seconds.

Setting a ServerAliveInterval value less than the ClientAliveInterval interval
forces the client to send a Null message to the server.  Hence, avoiding the
timeout.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
Co-authored-by: Shoaib Lari <slari@pivotal.io>