CI: Add `bin_gpdb_[platform]_rc` as Release_Candidate job output
A gate job is added for Release Candidate to make sure that all the
release candidate jobs passed for gpdb_src and bin_gpdb for centos6,
centos7 and sles11 platform.

The Release_Candidate job verifies that the commit SHA of gpdb_src and
all the bin_gpdb resources are the same. If the versions don't match,
the job will fail.

The bin_gpdb_[platform]_rc resources are put to a stable builds bucket
so that they can be consumed by integration and components pipelines

Co-authored-by: Alexandra Wang <lewang@pivotal.io>
Co-authored-by: Kris Macoskey <kmacoskey@pivotal.io>
Co-authored-by: Trevor Yacovone <tyacovone@pivotal.io>