Fix CScalarIdent's PopCopyWithRemappedColumns logic
When we create a copy of an operator using remapped columns the mapped
columns should align with whatever kind it is in the origical copy. That
means if a column was a table column, the copy should also refer the new
column as a table column and if it was a computed column it should be a
computed column.

For the CScalarIdent, the columns are blindly created as a computed
column in `PopCopyWithRemappedColumns`. This commit fixes the issue.

Below is a query that exposes the issue.

```
CREATE TABLE foo (a int);

CREATE TABLE bar (b int);

CREATE INDEX idx1 ON bar USING btree (b);

explain WITH 	x AS ( SELECT a FROM foo UNION ALL SELECT a FROM foo ),
		y AS ( SELECT a, 1 AS cust_type FROM x)
SELECT a FROM y WHERE (a IN (SELECT b FROM bar WHERE b = 2));
```

The predicate `(a IN (SELECT b FROM bar WHERE b = 2)` is pushed down to
the children of `UNION ALL` and when `UNION ALL` have `CLogicalIndexGet`
as one of its children the entire predicate is copied with columns
being remapped. When it comes to remapping the `CScalarIdent`, it
blindly creates a computed column while the column in `CScalarIdent`
refers to a table column.

When we query for stats of copied and pushed down `CLogicalIndexGet`, the
new computed column ends up being in the used columns for the index.
However since it is marked as a computed column, there is no mapping to
a real column, and therefore no stats exits and it is both logically and
technically wrong.