Make gprecoverseg use new pg_backup flag --force-overwrite
This is needed during gprecoverseg full to preserve important files
such as pg_log files. We pass this flag down the call stack to prevent
other utilities such as gpinitstandby or gpaddmirror from using the
new flag. The new flag can be dangerous if not used properly and
should only be used when data directory file preservation is
necessary.