Resolve GPDB_91_MERGE_FIXME about transaction_deferrable.
gp_dispatch = false on utility command is correct, as cannot dispatch the SET
command yet as not established the transaction yet on
QEs. transaction_deferrable is only useful with serializable isolation level as
per upstream docs, so added note to start dispatching the same when we support
serializable isolation level.