Remove deduplication in hyperloglog code
We had significant deduplication in hyperloglog extension and utility
library that we use in the analyze related code. This commit removes the
deduplication as well as significant amount dead code. It also fixes
some compiler warnings and some coverity issues.

This commit also puts the hyperloglog functions in a separate schema
which is non-modifiable by non superusers.

Signed-off-by: Bhuvnesh Chaudhary <bchaudhary@pivotal.io>