sles11 compile fix; sync_tools refactor
This is a common issue with sles11 - that it doesn't
currently include support for > TLS v1.1. Many upstream endpoints over
the last 6 months have started to enforce > TLS v1.1.

We resolved this by separating the sync_tools call into a separate task,
such that we could run this task from a centos docker image, prior to
compiling on the correct OS.

These changes were backported to other compile jobs. We are pushing this
change to resolve the sles11 blocker, but we are still experiencing
difficulty with windows.

Co-authored-by: Lisa Oakley <loakley@pivotal.io>
Co-authored-by: Trevor Yacovone <tyacovone@pivotal.io>
Co-authored-by: Ed Espino <edespino@pivotal.io>