Fix incremental analyze for non-matching attnums
To merge stats in incremental analyze for root partition, we use leaf
tables' statistics. In commit b28d02971fd85977f7d0812cc73db935288febaf,
we fixed an issue where child attnum do not match with a root table's
attnum for the same column. After we fixed that issue with a test, that
test also exposed the bug in analyze code.

This commit fixes the issue in analyze using the similar fix in
b28d02971fd85977f7d0812cc73db935288febaf.