Fix psql \db command for Greenplum 6 servers
Commit 4483b7d3bc16fd8b78076543f1ae696d267944b7 remove spclocation
from the tablespace catalog, but the \db command in psql wasn't
updated to match the corresponding Greenplum version as it was
backported prior to when it was introduced in upstream. This will
eventually go away as we merge with PostgreSQL, but that's not an
excuse for not fixing what is broken.

Discussion: https://github.com/greenplum-db/gpdb/pull/5238
Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>