Do not select the index if first key of a multi key index is not covered
by the predicate, because such index scan will be expensive.
If a multi-key index has been selected to be a potential match for given
predicate, then make sure that first key of the index is present in the
predicate.
For example, with schema: t(a, b , c, d), index i(b ,a, c)
and predicate: c = 10

No index path will be chosen since the given predicate does not cover
first key of the index i.

However if the predicate is (b = 1 and c = 10) then this index can be
selected.

Ideally, two alternatives should be generated in this case,
1. Table scan with c = 10 as filter
2. Index scan with index i (bad index choice)

And cost model should penalize the bad index scans and make the
decision. Currently, our cost model does not cost index scans used for
predicates that skip the first index keys, differently, than an index
scan used for a predicate that covers the first key of the multi-key
index.

With cost model equipped to handle such index paths, we can generate
such paths and leave it up to the costing. But for now, disallow them.

Also fix the tests and add new tests to CBtimapTest.

Signed-off-by: Dhanashree Kashid <dkashid@pivotal.io>