Get sequence code closer to upstream.
This patch attempts to get the sequence code as close to upstream as
possible. There is still mix of code from various upstream versions due to past
cherry-picks for bug fixes so its hard to state which exact upstream version it
matches but atleast all the gpdb specific modifications are minimized and
unnecessary code movement has been removed.