Let dispatcher raise error when a DTM broadcast fails
There already is a PG_TRY() ... PG_CATCH() block to handle such errors
in the DTM retry logic.  This change makes it easier to test the error
handling in the retry logic.

The patch also fixes a bug in dispatcher that invoked CopyErrorData
interface incorrectly, without switching to a memory context other that
ErrorContext.