Test case for DTM retry error handling
The test hits the PG_CATCH() block in DTM retry logic.  It uncovers a
bug in that part of the code, leading to PANIC due to
ERRORDATA_STACK_SIZE exceeded.

The upper limit on dtx_phase2_retry_count is increased to 15.  That
allows to keep the test simpler by avoiding PANIC due to max retry count
reached.