tuptoaster: gracefully handle changes to TOAST_MAX_CHUNK_SIZE
GPDB 4.3 uses a slightly smaller TOAST_MAX_CHUNK_SIZE, and we claim to
be able to upgrade those toast tables, but there's currently no support
in tuptoaster:

	ERROR:  unexpected chunk size 8138 (expected 8140) in chunk 0 of 4
	for toast value 16486 in pg_toast_16479 (tuptoaster.c:1840)

Add some tests to flush this out. This adds the
gp_test_toast_max_chunk_size_override GUC, which allows a superuser to
manually reduce the chunk size used by toast_save_datum().

We can handle changes to TOAST_MAX_CHUNK_SIZE by taking a look at the
size of the first chunk. For the full toast_fetch_datum(), we can
optimistically assume that the max chunk size is TOAST_MAX_CHUNK_SIZE,
and then adjust if that turns out to be false. For the random access in
toast_fetch_datum_slice(), though, I haven't tried a similar
optimization -- if we don't know the chunk size to begin with, it seems
like there are too many corner cases to keep track of when jumping into
the middle of a toast relation. So toast_fetch_datum_slice() will always
open the first chunk, which may somewhat impact performance for some
queries.