Adding a filter for the GiST B-tree path
Most GiST indexes are lossy indexes. When GiST follows the Bitmap Index
Path, this fact is not an issue since Bitmap indexes are also lossy and
there is already a recheck condition that gets added to the plan.
However, for the B-Tree Index Path, we must account for this as B-Tree
indexes are not lossy.

Currently, if we are very conservative and we re-add all of the index
quals to the residuals so that they can later be added to the plan as a
filter on top of the index condition. There are some cases where this
extra filter is not necessary.

For example:
- GiST indexes on type box are not lossy.
- PostGis queries that have a function in the WHERE clause often do not need
this filter on top as GPORCA creates a predicate as the index condition
and will use the given function from the WHERE clause as a filter.
```
EXPLAIN SELECT * FROM o_table WHERE ST_Contains(geom, ST_Point(1, 1));
                                                                          QUERY PLAN
---------------------------------------------------------------------------------------------------------------------------------------------------------------
 Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..790699.49 rows=281102 width=36)
   ->  Index Scan using rtree_table_o on o_table  (cost=0.00..790661.78 rows=93701 width=36)
         Index Cond: geom && '0101000000000000000000F03F000000000000F03F'::geometry
         Filter: _st_contains(geom, '0101000000000000000000F03F000000000000F03F'::geometry) AND geom && '0101000000000000000000F03F000000000000F03F'::geometry
```

Signed-off-by: Ekta Khanna <ekhanna@pivotal.io>