Disable GiST index expressions and partial indexes
Since index expressions are not supported for the Bitmap Index Path,
GiST also does not support this option. We extended this for all of
GiST, so that even if a GiST index follows the B-Tree path, it will not
create an alternative if the query has an index expression. For example,
given a table foo with a GiST index on p, the query `SELECT * FROM foo
WHERE p IS NULL` will not create a plan containing the GiST index.

For partial indexes, since we cannot obtain all of the information
regarding the storage type of the different partitions, we cannot make
an informed decision regarding the use of GiST indexes. Therefore, we
currently do not support partial GiST indexes.

Signed-off-by: Ekta Khanna <ekhanna@pivotal.io>