Unbreak compilation of SNMP support on macOS
Commit 3b3adb2be3c561fd7638bed96b859ee2e408f3a2 most likely broke
compilation of the SNMP related code on macOS, as it required a
feature test macro flag to enable the right bits in <sys/types.h>.
There however seems to be little to no reason for the feature test
macros to be defined here either, so remove and try to unbreak the
build rather than resurrect.

Also enable snmp support in Travis CI compilation so we get a heads
up the next time it gets broken.