Replace Oid preallocation with pg_upgrade Oid handling
Previously, the Greenplum pg_upgrade was syncronizing the Oids across
upgrades using the backend Oid pre-allocation. All pre-allocations
were generated in an Oid file, which was executed before the schema
was restored. When doing parallel restores, this strategy becomes
problematic as the Oid preallocation may be done in a backend other
than the one where the Oid is required. This commit rips out the use
of Oid preallocation in favor of using the upstream Oid handling in
pg_upgrade where the Oids are set just before being required. A major
difference is that we use the Oid preassignment scaffolding under the
covers, but in a more upstream merge friendly way.

For now, a known regression is that toast tables which in the old
cluster doesn't match the naming convention in the new cluster fails
to restore. Drop before running test until we know how to handle these.