Improved handling of empty lines and comments in commands log
There are a lot of blank lines in the 'gpinitsystem' log. These are
read directly from the configuration file using commands 'cat'. The
comments are turned into blank lines, but they are not removed:

  gpinitsystem:node:gp6-[INFO]:-Start Main
  gpinitsystem:node:gp6-[INFO]:-Command line options passed to utility = -c ../gpinitsystem_config
  gpinitsystem:node:gp6-[INFO]:-Start Function CHK_GPDB_ID
  ...
  gpinitsystem:node:gp6-[INFO]:-End Function CHK_FILE
  gpinitsystem:node:gp6-[INFO]:-Dumping gpinitsystem_config to logfile for reference

  ARRAY_NAME="EMC Greenplum DW"

  SEG_PREFIX=gpseg

  PORT_BASE=40300

This extends the exclusion regex used when appending to the logfile
to remove blank lines completely and to handle comments that doesn't
start on column zero.

Co-authored-by: Daniel Gustafsson <dgustafsson@pivotal.io>
Reviewed-by: Daniel Gustafsson <dgustafsson@pivotal.io>