Fix expected output issues in mbregress.sh.
I don't think anyone's tried to run this on Greenplum, until @imshubhangi
reported in github issue #5241, that it doesn't work. Most of the failures
are just a matter of updating the expected output for extra NOTICES, and
using gpdiff.pl to mask out differences in row order. That's what this
patch does.

However, there is one remaining failure, in the COPY TO STDOUT at the
end of the 'sjis' test, which looks like a genuine bug. That needs to be
investigated and fixed, but let's do that as a separate commit.