Backport upstream fix for bug in tuplesort_skiptuples().
We had backported the tuplesort_skiptuples() function, as part of commit
fd6212ce130 which backported upstream support for ordered-set aggregates.
Since we backported the feature, we also need to keep backporting all
bugfixes that follow, until we catch up to that in the merge.

In GPDB, the same fix needs to also be applied to the "mk sort" variant.

Fixes github issue #5334.

Upstream commit:

commit 1def747db68b234b54b784c74dfbe329f5478b79
Author: Tom Lane <tgl@sss.pgh.pa.us>
Date:   Tue Dec 24 17:13:02 2013 -0500

    Fix inadequately-tested code path in tuplesort_skiptuples().

    Per report from Jeff Davis.