Fix INSERT RETURNING on partitioned table.
The ResultRelInfos we build for the partitions, in slot_get_partition(),
don't contain the ProjectInfo needed to execute RETURNING. We need to
look that up in the parent ResultRelInfo, and when executing it, be
careful to use the "parent" version of the tuple, the one before
mapping the columns for the target partition.

Fixes github issue #4735.