Give proper error on DELETE RETURNING on AO table.
We don't support this at the moment. Would be nice to make it work, but
in the meanwhile, let's at least give a proper error message, instead of
attempting to read the AO table as if it was a heap table.

There was a FIXME comment about this, and it was also reported in github
issue #4735.