pg_dump: improve speed of partition handling
pg_partitions is a massively complicated view, and here we are using it
just to grab the relnames of the partition/subpartition tables so that
we can join them back to pg_class and get their OIDs. The
pg_partition/_rule catalog tables already have the OIDs; there's no
reason to use the view. Perform the join here manually and speed up the
process.

This change is motivated by a serious performance regression when
dumping the regression database schema -- this query was taking about
half a second to execute, and EXPLAIN showed about a dozen nested loop
joins. The new query is nice and compact.

Co-authored-by: Jim Doty <jdoty@pivotal.io>