Fix parse analysis of PARTITION BY clause containing NULLs.
NULLs are not allowed in the PARTITION START or END clauses, but the checks
for NULLs were performed only after sorting the list of partitions. The
sort function did not consider the possibility of NULLs, and would therefore
pass NULLs to comparison functions, even if they're marked as strict. That
could cause a segfault.

Fixes https://github.com/greenplum-db/gpdb/issues/5503