Ensure that `SUBPARTITION TEMPLATE` only appears after `SUBPARTITION BY`
This commit changed the way the grammar was parsing the subpartition
information to ensure that the templates could not exist without a
`SUBPARTITION BY`.

Testing coverage for the case where the `SUBPARTITION TEMPLATE`
expression is written before a `SUBPARTITION BY` was added.

The error displayed when the template appeared after a Partition was
changed to point to the TEMPLATE

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>
Co-authored-by: Jesse Zhang <sbjesse@gmail.com>
Co-authored-by: Adam Berlin <aberlin@pivotal.io>