resgroup: restore dir permission in unittest.
In unittest test_when_cpuset_gpdb_dir_bad_permission we check the case
that cpuset dir has bad permission, after the test we need to restore
the permission otherwise the cleanup process will fail to remove the
dir.  Previously we only do the restore when gpdb version >= 6.0.0 so
the case will fail on 5X.  Fixed by restoring on all versions.