Revert "Revert "Avoids deadlocks for concurrent index creation on AO tables.""
This reverts commit c3726b0936b25d7d43085e10e7d188c848bfac3d.

The original commit, which we're re-introducing with this Revert, was failing
because some isolation2 tests had this situation:

Session one successfully reindexes with a blocking call without committing,
continuing to hold the lock. Session two attempts to create an index on the
same table, which is locked by session one.

We introduced backgrounding of session two, which allows session one to return
and commit, releasing the lock. Our change enables the test runner to behave in
a realistic concurrent way, after the introduction of the
ShareRowExclusiveLock.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>