Fix EXPLAIN timing off option.
This chunk of code was left out during the 9.2 merge. As a result, "EXPLAIN
(analyze, timing off) still collected and printed the timing information.
In other words, the "timing off" had no effect.

No test added, because EXPLAIN ANALYZE is a bit tricky to deal with in
regression testing. I think gpdiff.pl would mask out this difference in
the EXPLAIN output. Or if not, then the extra memory information and
"Total runtime" that's printed in EXPLAIN ANALYZE would cause spurious
failures. The code is now pretty much identical to upstream, so I'm not too
worried about the lack of coverage.