Add tests ensuring correct handling of Full and left outer joins
1. Add a test for a full outer join query on varchar columns
In such scenario, planner expects a relabletype node on top of varchar
column while looking up for a Sort operator. Please refer commit fab435e
for more details.  Add a test for such queries and disable hashjoin to
make sure that a planner is able to generate a plan with merge join
successfully.

2. Add a test for a query with an Agg and left outer join
This test is to ensure that ORCA produces correct results, by performing
a two stage aggregation on top of a co-located join. Corresponding plan
test has been added in the ORCA test suite.