Refactor parse analysis of DISTRIBUTED BY, to fix bug with inheritance.
Previously, we would create a GpPolicy object to represent the DISTRIBUTED
BY clause in the parse analysis stage, and store it in CreateStmt->policy.
That's problematic with table inheritance, however. A GpPolicy refers to
the columns by attribute numbers, and we don't know the final attribute
numbers of the columns at parse analysis yet, because the table might
inherit additional columns from a parent table. We tried to adjust the
attribute numbers in MergeAttributes(), but that was complicated, and
we got it wrong in the case reported in github issue #4462.

To fix, delay constructing the GpPolicy object until after we have
processed table inheritance. transformDistributedBy() now returns
a DistributedBy object, which is equivalent to the GpPolicy object we
used to create there, but it refers to the columns by name rather than
attribute number.

With CREATE TABLE AS, we still create a GpPolicy to represent the
DISTRIBUTED BY clause, if the user gave one explicitly. That works,
because you cannot create an inherited table with CREATE TABLE AS.
This is slightly inconsistent with the regular CREATE TABLE case, but
the processing is quite different anyway. In the CTAS case, we deduce
the default policy in the planner, rather than parse analysis, and
we deal with a GpPolicy from that point on.

Fixes https://github.com/greenplum-db/gpdb/issues/4462