Remove obsolete backwards-compatibility code in hba parsing.
This chunk was introduced back in 2009. The commit message said:
(from the old git repository, long before GPDB was open sourced)

    Some backwards compatability fixes.
    Allow database name to come before arguments in psql
    Allow old pg_hba syntax " local all userid indent sameuser" as well as new syntax  of " local sameuser userid ident".

Back then, there was a comment in the code too, that said:

    XXX: attempt to do some backwards compatible parsing here?

That commit was removed in 2010, when the hba.c code was synced with
PostgreSQL 9.0.

I don't think we need to maintain compatibility with ancient pg_hba.conf
files anymore. If anyone still has a config file with those fields swapped,
it's time to fix the file.