Fix rechecking index quals in with lossy index operators on AO tables.
In 8.4, a 'recheck' flag was added to TBMIterateResult. But the bitmap
AO table scan methods didn't get the memo. Because of that, they would
fail to recheck the original index quals, when the index operator was
lossy. Test case included to demonstrate the bug.

Unrelatedly, remove a FIXME comment about the need to have EvalPlanQual
recheck functions for bitmap AO table scans. That 'recheck' method is
in fact unnecessary for AO tables, as explained in the comment that
replaces the FIXME. This is confusing, so pay attention: the 'recheck'
functionality that this commit fixes, is unrelated to the FIXME comment
and the Recheck "access methods". But I found the bug with the
TBMIterateResult.recheck field when I started to look into that. I'm glad
I was confused about that myself, or I would not have tested the thing
that was actually broken :-)

As a side-note, it's confusing that we have essentially duplicated the
bitmap heap/AO/AOCS scan code so many times. We have Bitmap Table Scan for
heap, AO and AOCS tables, when ORCA is used. And separate Bitmap Heap Scan
and Bitmap AO/AOCS Scan nodes when the Postgres planner is used. We
should refactor that, but not in this patch.