Avoid calling GPDB backend functions directly from ORCA translator.
It was safe in this case, because numeric_is_nan() doesn't throw an
error, but let's follow the usual convention and go through a wrapper
function gpdbwrappers.cpp. For the sake of consistency, and safety in
the unlikely case that numeric_is_nan() is modified in the future to
do more than a simple comparison.