Fix non-text format EXPLAIN of GPDB-specific plan nodes.
This was a bug-of-omission when the upstream support for XML and JSON
formats was merged. The initialization of the 'sname' variable was added
for every node type handled in the switch-case statement, but the GPDB
addons didn't get the same treatment. As a result, non-text format
EXPLAIN of one of these GPDB-specific nodes produced garbage output, or
a segfault.