Don't emit extra ALTER FUNCTION OWNER TO commands after CREATE LANGUAGE.
In pg_dump, we had a GPDB-specific hack to change the owners of the
handler and validator functions of the language, after dumping the CREATE
LANGUAGE command. This was added back in 2008, with commit message:

    MPP-3774: Change pg_dump to emit ALTER FUNCTION ... OWNER statements
    for functions underlying created languages for languages defined
    using pg_pltemplate-specified languages.

I looked into MPP-3774, but I couldn't make much sense of it.

Title:
    pg_dumpall doesn't dump Language function owners
Description:
    pg_dumpall doesn't dump the permissions for the CREATE LANGUAGE. need to
    modify the function owners. so ALTER FUNCTION ... OWNER TO ...

PostgreSQL never had that code. I don't think it is relevant anymore,
there's been a lot of change in the way languages are handled. Firstly,
they are usually created as part of extensions, these days, so the
"language template" mechanism is obsolete. Secondly, when this change was
made in GPDB, languages didn't have owners. That was added later.

All in all, I don't think this is needed anymore, so remove it, and dump
languages and their functions just like in the upstream.

PS. If this was still valid and needed, I think we'd also need to emit the
ALTER FUNCTION commands for the 'laninline' function.