Clarify comments around const-evaluation of Params.
GPDB has a concept of "one-off" plans. In PostgreSQL, the planner doesn't
normally const-evaluate stable functions, but GPDB does that. Such plans
cannot be reused across invocations, so they're marked as "one-off".

Before the PostgreSQL 9.2 merge, the same "one-off" flag was used to mark
plans where query Params were aggressively const-evaluated, for similar
reasons. That mechanism was obsoleted in 9.2, however, when the concept
of "custom" and "generic" plans was introduced. The planner will now try
to create "custom" plans for the specific query parameters, without any
GPDB adde code.

The snippet in In eval_const_expressions that this removes was a no-op.
There used to be an extra condition, before the 9.2 merge, so that it
const-evaluated Params more aggressively. But that was removed as part
of the merge, and as a result, the removed if()'s condition was always
false. We don't need the more aggressive stance here, because the upstream
custom plans mechanism does the same. Move the comment explaining the
aggressive const-evaluation of stable functions to evaluate_function(),
that seems like a better place for it, anyway.

Remove the GPDB_92_MERGE_FIXME comment in BuildCachedPlan(). I'm not sure
why we had a check for boundParams there before. AFAICS, marking the plans
as one-off in eval_const_expressions_mutator (the code that was removed),
should've been enough, and the check for boundParams in plancache.c was
redundant (and sub-optimal, because some plans were marked not-for-reuse,
which could've been reused safely). In any case, there isn't anything
special about query parameters in this code anymore, it's all upstream
code, so we don't need an extra boundParams check either.

Add a test case for the const-evaluation of query Params, too. It was
passing before these changes as well, but seems good to have a test for it.